+++ app/src/packages/work/knowledge/index.config.ts
@@ -1,4 +1,5 @@
 export default {
   navigationStyle: 'custom',
   navigationBarTitleText: '我的资料库',
+  enablePullDownRefresh: true,
 };